+++ projects/demo/src/app/custom-instance/custom-instance.service.ts
@@ -1,0 +1,21 @@
+import { Injectable } from '@angular/core';
+import { CustomNGXLoggerService, NGXLogger } from '../../../../../src/public_api';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class CustomInstanceService {
+  private logger: NGXLogger;
+
+  constructor(
+    customNgxLoggerService: CustomNGXLoggerService,
+  ) {
+    this.logger = customNgxLoggerService.getNewInstance({ partialConfig: { context: 'CustomInstanceService' } });
+  }
+
+  logWithContext(): void {
+    this.logger.error('Logging from CustomInstanceService');
+  }
+}
+
+