+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - Merged features waiting to be published in upcoming version
 
 
+## [5.0.11] - 2022-05-24
+### Added
+- Added INGXLogger.partialUpdateConfig that updates only some fields of the config
+- Added partialConfig parameter to CustomNGXLoggerService.getNewInstance
+- Added context to INGXLoggerConfig, context is written in {} to the log for every log messages
+
+
 ## [5.0.10] - 2022-05-24
 ### Added
 - Added getNewInstance in CustomNGXLoggerService that allows more params for instantiating a new instance of NGXlogger